smarter tx matching in BitBay 3.0 API importer
We can’t always just take next 3 lines from the list and match
them, because some transactions happening at the exact same timestamp
get their lines mixed up sometimes.

But we also can’t just collect all lines with a given timestamp and
match them, because sometimes there might be a 1 (or more?) second
difference between two lines of a single transaction.

So the previous approach was to wait until an incoming transaction has
a significant enough time difference from the ones collected, and then
match all those before it. But this sometimes collected too many lines
together if they were all 1 second after another… and it could still
crash sometimes.

So now it always tries to match each new set of 3 (or 6, 9…)
transactions as soon as possible, but only if it can, and otherwise it
collects them until a proper set forms.